+++ tooling/template-base/server/index.js
@@ -12,7 +12,41 @@ const heliumEndpoint = process.env.HELIUM_ENDPOINT;
 startServer();
 
 async function startServer() {
+  if (!heliumEndpoint) {
+    throw new Error(`
+    HELIUM_ENDPOINT environment variable is not set.
+    The development server should be started 'helium dev'
+    `);
+  }
+
   const app = express();
+
+  if (!isProduction) {
+    const expressPlayground = require('graphql-playground-middleware-express').default;
+    const fetch = require('isomorphic-unfetch');
+
+    app.use(express.json());
+
+    app.get('/graphiql', expressPlayground({ endpoint: '/graphql' }));
+    // proxy GraphQL Playground's requests because of CORS errors
+    app.post('/graphql', async (req, res) => {
+      const options = {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(req.body)
+      };
+
+      fetch(heliumEndpoint, options)
+        .then(tiRes => {
+          res.status(tiRes.status);
+          return tiRes;
+        })
+        .then(r => r.json())
+        .then(data => {
+          res.send(data);
+        });
+    });
+  }
 
   let viteDevServer;
   if (isProduction) {